Added new save service for coin, reworked bestscore service, added new UI, code refactoring

+++ RunnerGame/Assets/CodeBase/Data/BestScore.cs
@@ -2,6 +2,6 @@
 {
     public struct BestScore
     {
-        public int Score;
+        public float Score;
     }
 }
