+++ RunnerGame/Assets/CodeBase/Services/BestScore/ISaveTheBestScore.cs
@@ -2,7 +2,7 @@ namespace CodeBase.Services.BestScore
 {
     public interface ISaveTheBestScore : IService
     {
-        int TheBestScore { get; }
+        float TheBestScore { get; }
         void Save();
         void Load();
     }
